Clean up test dependencies (#1858)
