+++ tests/common/gcp_api/cloudsql_test.py
@@ -21,7 +21,7 @@ from google.oauth2 import credentials
 
 from tests import unittest_utils
 from tests.common.gcp_api.test_data import http_mocks
-from tests.common.gcp_type.test_data import fake_cloudsql
+from tests.common.gcp_api.test_data import fake_cloudsql_responses as fake_cloudsql
 from google.cloud.forseti.common.gcp_api import cloudsql
 from google.cloud.forseti.common.gcp_api import errors as api_errors
